backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.4
pydantic==2.10.4
